Ajustado o parametro de percentual de reajust para double

+++ src/main/java/br/com/gar/padaria/controllers/ProdutoController.java
@@ -36,7 +36,7 @@ public class ProdutoController {
     }
 
     @GetMapping("reajustarPreco/{percentualReajuste}")
-    public int reajustarPreco(@PathVariable("percentualReajuste") Integer percentualReajuste) {
+    public int reajustarPreco(@PathVariable("percentualReajuste") Double percentualReajuste) {
         return produtoRepository.f_reajuste_preco_venda(percentualReajuste);
     }
